+++ CHANGELOG.md
@@ -3,6 +3,8 @@ v0.3.0 (in development)
 - Table cell values that begin with a combining character will now have a space
   prepended
 - Added a `padding` option
+- Strings with indeterminate widths now cause an `IndeterminateWidthError` to
+  be raised
 
 v0.2.0 (2018-06-16)
 -------------------